+++ core/components/romanesco/elements/plugins/07_computations/c_settings/updatestyling.plugin.php
@@ -96,14 +96,40 @@ switch($eventName) {
 
         // Regenerate styling elements if theme settings were updated or deleted
         if ($updatedSettings || $deletedSettings) {
-            // Clear cache, to ensure build process uses the latest values
-            $modx->getCacheManager()->delete('clientconfig',array(xPDO::OPT_CACHE_KEY => 'system_settings'));
-            if ($modx->getOption('clientconfig.clear_cache', null, true)) {
-                $modx->getCacheManager()->delete('',array(xPDO::OPT_CACHE_KEY => 'resource'));
+
+            // Set theme.variables path for current context
+            $themesFolder = $modx->getOption('base_path') . 'assets/semantic/src/themes/';
+            if ($currentContext) {
+                $themeVariablesPath = $themesFolder . $currentContext . '/globals/theme.variables';
+            } else {
+                $themeVariablesPath = $themesFolder . 'project/globals/theme.variables';
             }
 
-            // Terminate any existing gulp processes
-            $killCommand = "ps aux | grep '[g]ulpfile " . $modx->getOption('assets_path') . "semantic/gulpfile.js' | awk '{print $2}'";
+            // Clear cache, to ensure build process uses the latest values
+            $modx->getCacheManager()->delete('clientconfig',array(xPDO::OPT_CACHE_KEY => 'system_settings'));
+
+            // Grab variables after cache rebuild
+            $newSettings = $clientConfig->getSettings($currentContext);
+
+            // NB: for some reason, using getChunk directly here doesn't work (when using settings array as content)
+            $themeVariables = $modx->getObject('modChunk', array('name'=>'themeVariables'));
+            $themeVariables->setCacheable(false);
+
+            // Write to theme.variables
+            $pathInfo = pathinfo($themeVariablesPath);
+            $path = $pathInfo['dirname'];
+
+            if (!file_exists($path)) {
+                mkdir($path, 0755, true);
+            }
+
+            file_put_contents(
+                $themeVariablesPath,
+                $themeVariables->process($newSettings, $themeVariables)
+            );
+
+            // Terminate any existing gulp processes first
+            $killCommand = "ps aux | grep '[g]ulp build-' | awk '{print $2}'";
             exec(
                 'kill $(' . $killCommand . ') 2> /dev/null',
                 $output,